* Fix cache clear.
