+++ app.js
@@ -2,7 +2,7 @@ const express = require('express')
 const app = express()
 const router = require('./routes/index')
 const bodyParser = require('body-parser')
-const { handleCachedData } = require('./middleware/cache')
+const { handleRequestsData } = require('./middleware/cache')
 const path = require('path')
 const config = require('config')
 const { authenticateJwt } = require('./middleware/jwtAuthentication')
@@ -11,7 +11,7 @@ app.use(express.static(path.join(__dirname, 'client', `${config.static}`)))
 app.use(bodyParser.urlencoded({ extended: false }))
 app.use(bodyParser.json())
 app.use(authenticateJwt)
-app.use(handleCachedData)
+app.use(handleRequestsData)
 app.use(router)
 
 module.exports = app
